Eliminate compile time dependency on runtime-language-epb
