+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/interop/generic/ReadArrayElementNode.java
@@ -7,8 +7,7 @@ import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.profiles.BranchProfile;
 import org.enso.interpreter.Constants;
 import org.enso.interpreter.dsl.BuiltinMethod;
-import org.enso.interpreter.epb.node.CoercePrimitiveNode;
-import org.enso.interpreter.node.expression.foreign.CoerceNothing;
+import org.enso.interpreter.node.expression.builtin.interop.syntax.HostValueToEnsoNode;
 import org.enso.interpreter.runtime.EnsoContext;
 import org.enso.interpreter.runtime.builtin.Builtins;
 import org.enso.interpreter.runtime.error.PanicException;
@@ -22,13 +21,12 @@ public class ReadArrayElementNode extends Node {
   private @Child InteropLibrary library =
       InteropLibrary.getFactory().createDispatched(Constants.CacheSizes.BUILTIN_INTEROP_DISPATCH);
 
-  private @Child CoercePrimitiveNode coercion = CoercePrimitiveNode.build();
-  private @Child CoerceNothing nothingCoercion = CoerceNothing.build();
+  private @Child HostValueToEnsoNode toEnso = HostValueToEnsoNode.build();
   private final BranchProfile err = BranchProfile.create();
 
   Object execute(Object array, long index) {
     try {
-      return nothingCoercion.execute(coercion.execute(library.readArrayElement(array, index)));
+      return toEnso.execute(library.readArrayElement(array, index));
     } catch (UnsupportedMessageException e) {
       err.enter();
       Builtins builtins = EnsoContext.get(this).getBuiltins();
